Update policy create UI for common policies and API specific policies
